removed a console.log from auth.js
added data object of req.user in about handlebars, homepage handlebars, muscles and muscle-specific handlebars,

added a console.log in the user-page.js as a test,

added a console.log to the logout page,

a lot of red, but all i did was hit alt + shift + f and so reformating the page for proper indentation makes it look like i deleted everything lol

Anyways, I did change the variable passed into the if/else statements: one in the nav-bar to make one button or the other show (technically an <a>), the other at the bottom I am trying to make the script logout.js fire but couldn't get it to work.

We still need logout code.